Update to RxSwift 6
