--- Cartfile
@@ -1,2 +1,0 @@
-github "ReactiveX/RxSwift" ~> 5.0
-github "kazuhiro4949/PagingKit" ~> 1.9.0
