+++ Example/RxPagingKit/AppDelegate.swift
