--- RxPagingKit.xcworkspace/contents.xcworkspacedata
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "group:RxPagingKit.xcodeproj">
-   </FileRef>
-   <FileRef
-      location = "group:iOS Sample/iOS Sample.xcodeproj">
-   </FileRef>
-</Workspace>
