+++ _Pods.xcodeproj
@@ -1,0 +1,1 @@
+Example/Pods/Pods.xcodeproj
