--- iOS Sample/Cartfile.resolved
@@ -1,2 +1,0 @@
-github "ReactiveX/RxSwift" "5.0.1"
-github "kazuhiro4949/PagingKit" "1.10.0"
